Move DeviceAPI into its own header file
Nothing in Expr.h needs it, and at least one of the few users (Target.h) didn't need anything else in Expr.h.
