+++ src/CMakeLists.txt
@@ -331,6 +331,7 @@ set(HEADER_FILES
   Deinterleave.h
   Derivative.h
   DerivativeUtils.h
+  DeviceAPI.h
   DeviceArgument.h
   DeviceInterface.h
   Dimension.h
